+++ plugin.py
@@ -1,4 +1,5 @@
 import sublime
+import sublime_plugin
 
 def plugin_loaded():
     disable_native_php_package()
@@ -41,3 +42,42 @@ def disable_native_php_package():
         ignored_packages.sort()
         settings.set('ignored_packages', ignored_packages)
         sublime.save_settings('Preferences.sublime-settings')
+
+def is_finish_completion_context(view):
+    for s in view.sel():
+        if view.substr(s.end()) != ')':
+            return False
+    return True
+
+class FinishCompletionContext(sublime_plugin.EventListener):
+    def on_query_context(self, view, key, operator, operand, match_all):
+        if key == 'finish_completion':
+            if is_finish_completion_context(view):
+                return True
+
+        return None
+
+class FinishCompletionCommand(sublime_plugin.TextCommand):
+
+    """
+    Deletes selection, moves forward one character, and
+    optionally inserts specified characters after the
+    move forward.
+    """
+
+    def run(self, edit, insert_characters = None):
+        if not is_finish_completion_context(self.view):
+            return
+
+        for s in self.view.sel():
+            self.view.erase(edit, s)
+
+        self.view.run_command('move', {
+            'by': 'characters',
+            'forward': True
+        })
+
+        if insert_characters != None:
+            self.view.run_command('insert', {
+                'characters': insert_characters
+            })