+++ CHANGELOG.md
@@ -8,6 +8,7 @@ I know that many can't update to the latest build. The package will continue wor
 
 * Added: punctuation.definition.(end|begin).php scope to parens
 * Added: constant.character.format-argument scope for sprintf %s arguments in strings
+* Added: macro to finish completion by pressing backspace, ), or ;
 * Removed: ambiguous filetypes like install, profile, etc
 
 ## 0.17.0